plato_if: Implement erase-abort
Erase abort aborts queued output when a full-screen erase is
placed into the queue. This improves responsiveness when typing
a bit ahead. The PLATO system does that on its internal queue,
so this attempts to get the same results for this queue outside
the host. The original PLATO system had no additional queuing
between the host and the terminal, so there is no contemporary
implementation to base this one on.